+++ 03-Nbody/src/Vector.cpp
@@ -1,6 +1,6 @@
 #include "Vector.h"
 
-vec3::vec3(double x, double y, double z) : x(x), y(y), z(z) {}
+vec3::vec3(float x, float y, float z) : x(x), y(y), z(z) {}
 
 vec3 vec3::operator+(const vec3 &other) const
 {
@@ -12,18 +12,18 @@ vec3 vec3::operator-(const vec3 &other) const
     return vec3(x - other.x, y - other.y, z - other.z);
 }
 
-vec3 vec3::operator*(double c) const
+vec3 vec3::operator*(float c) const
 {
     return vec3(c*x, c*y, c*z);
 }
 
-double vec3::norm() const 
+float vec3::norm() const 
 {
     return std::sqrt(x*x + y*y + z*z);
 }
 
 vec3 vec3::normalized() const 
 {
-    double n = norm();
+    float n = norm();
     return vec3(x/n, y/n, z/n);
 }
